[FIX] Move coauthorships cell to core and make m-card use it.
