+++ decidim-core/app/cells/decidim/card_m/header.erb
@@ -14,4 +14,5 @@
   <% end %>
 
   <%= render :author if has_author? %>
+  <%= cell("decidim/coauthorships", model) if coauthorable? %>
 </div>
