+++ decidim-proposals/app/views/decidim/proposals/proposals/show.html.erb
@@ -7,7 +7,7 @@
 <%= render partial: "voting_rules" %>
 <div class="row column view-header">
   <h2 class="heading2"><%= @proposal.title %></h2>
-  <%= cell("decidim/proposals/coauthorships", @proposal) %>
+  <%= cell("decidim/coauthorships", @proposal) %>
 </div>
 <div class="row">
   <div class="columns section view-side mediumlarge-4 mediumlarge-push-8 large-3 large-push-9">
